stable/apply-stable-patches: assume sublevel as zero if not specified
In the class used for comparison of kernel versions, don't leave
sublevel as None if sublevel wasn't specified: it's optional and we mean
zero when not specifying it. Leaving as none may break version
comparison later.

While at it, make verbose mode to split when versions are specified for
commits being processed.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>