stable/sru-workflow-manager: fix version+abi match
When checking if dependent packages are built (-meta, -lbm, etc.), the
match for the version+abi is wrong: we were matching version-ABI*
instead of version-ABI.*, so for example if we are checking if
linux-meta 3.5.0-21 is built, 3.5.0-21* would be matched. We noticed the
bug in practice when linux-meta 3.5.0.21.* clashed with 3.5.0.216.15
omap4 -meta package.

Reported-by: Luis Henriques <luis.henriques@canonical.com>
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>