stable/sru-workflow-manager: always enable exact_match with getPublishedSources
We must do an exact_match, not a substring match (actually this is
confusing and I think exact_match should have been enabled by default...).
Otherwise we will get unwanted results for some packages which have the
same version in the ppa and search matches the substring, like when
searching for 'linux' source which will match linux and linux-signed
sources.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>