ktl/workflow.py: Add new devel_workflow database (dictionary)
This is the start of the work to make our stable tools to support
development kernel tracking bugs, in the same way as we handle kernel
stable updates. As the tasks on development tracking bugs differ
significantly from the ones needed on the stable tracking bugs, we
introduce a new dictionary for the task assignment. While at it,
document better both workflow dictionaries.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>