stable/sru-workflow-manager: process kernel-development-workflow bugs
With the new workflow for development kernels, process their bugs too.
Just factor the processing on a single function and pass the workflow we
want to process. Probably we will want to rename sru-workflow-manager to
something else, since this is now not only about SRU tracking bugs.

With the factorization, also move processing of URL to display in
bug_url to the function itself, no need to store the value on a global
place.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>