stable/sru-workflow-manager: set security tasks on derivative bugs properly
We agreed that if a master bug has publishing done to security pocket,
the derivative should also be published (as the derivative is likely a
rebase, or has also security fixes). This implements the proper checks
and also sets promote-to-security to Confirmed on the derivative if
master also has it set to Confirmed or Fix Released. I'm reusing the
existing check_master_task function, so adjust it and its usage for this
new check.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>