stable/sru-workflow-manager: be smarter when checking package builds
This is a huge change reestructuring the way we check for package
builds. Before, the bot was relying on tasks being set to an specific
status ('In Progress'), and only in this case we would check for builds.
Also, depending on the existence of the upload-to-ppa task, we were
using it also to conditionally check for the main package build on the
ckt ppa. Not only that, but in addition we were treating prepare-package-*
tasks as a dependency to the main prepare-package task, because we were
using only the prepare-package status to check if we needed to set the
promotion to proposed tasks forward.

Now, instead of treating all prepare-package tasks separately, before
we process bug tasks and actions we first check for the status of all
package builds, and depending on their status, we set the
prepare-package status right away to the proper state. Also, we look and
set what should be the proper assignee from information gathered from
launchpad, and if the packages were already released, we also go ahead
and set promotion to update/release tasks status and assignments.

Finally, upload-to-ppa task now only servers to indicate that a package
is ready to be uploaded from contents in git, for bugs where maintainer
and stable team split the work. The bot doesn't use it anymore for
anything.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>