ktl/workflow.py: disable upload-to-ppa task on armadaxp bugs
Since Ike himself prepares the packages and uploads them, there is no
reason to have an upload-to-ppa task on armadaxp tracking bugs.
upload-to-ppa is only useful to tell us, the stable team, that source
is ready and we can build the source packages and upload, after the
maintainer prepared and provided the source tree somewhere.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>