stable/sru-workflow-manager: be careful when changing status from Fix Released
When the copy of packages happen between pockets, we may be in an state
where packages are not available anywhere while the bot is running. This
means that we were misdetecting the package builds as not available.
Refactor the setting of build status to account for this, adding a
bug property that we use to wait before we go wild and reset the
prepare-package tasks if for some reason packages are really gone.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>