stable/sru-workflow-manager: don't set abi tasks to invalid when closing
When packages are published to the release or updates pockets, we can't
reliably detect anymore if an ABI bump was done, as we use the
difference between packages in ckt-ppa/proposed pockets and the ones in
the release/updates pockets (and don't check package publishing history,
which can be complex). Due to this, we can't just rely on this
detection when packages are released, otherwise we will set all
prepare-package tasks which are abi bumps to Invalid, even if they were
really done and released.

If packages are released and really we are unable to detect the abi
change, assume the state of the task is correct and don't change it.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>