ktl/tracking_bug.py: fix importance setting for nominations in devel series
Commit "ktl/tracking_bug.py: set development tasks to invalid"
unfortunately uncovered an weird launchpad behaviour when handling
development package tasks: since Ubuntu package tasks which are not a
nomination are considered the task for the development series, when you
nominate the development series in the bug, the status is now tracked
in the nomination ("Status tracked in <series>" as shown when opening
the bug in the browser).

The problem is if you try to set importance or any remaining fields in
the package task, and its status is tracked in a nomination, Launchpad
service will return an "HTTP Error 412: Precondition Failed", probably
because the importance and other fields are not "available" anymore. But
Launchpad API still provides those fields and there is no way to
differentiate when status is tracked in a nomination or not. So this
change tries to avoid the issue, if the series is a development
series we avoid messing with the importance field in the package task.
Also we properly set now the status to Invalid on it if this is an
stable tracking bug.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>