config-summary: fix negative choice handling in annotations merge
We were emmitting two entry for a negative annotation generated as a result
of a positive choice elsewhere in the case where that configuration option
already had an annotation.  The purpose of the merge is to end up with a
single entry for any one configuration.  Fix this up by merging the 'n'
into any existing annotation.

Signed-off-by: Andy Whitcroft <apw@canonical.com>