maint-startnewrelease - call fdr startnewrelease
Our process has changes so that we always bump ABI on releases.
This has been done to remedy a number of problems including
making sure that we always provide a backup means for booting
a working kernel. This logic is incorporated in the
startnewrelease target with the kernel, so let's use that
logic and not duplicate it in this script.

Signed-off-by: Steve Conklin <sconklin@canonical.com>