notify-stable-review: implement --sent-cover-letter=<Message-Id>
(Optional and unusual) Assumes the cover letter has already been sent
as <Message-Id>, so passes that to git-send-email as the --in-reply-to
to be used (and does not generate another cover letter).

Useful for "restarting" a busted notify-stable-review run, or for
attaching the review patch emails to a manually-sent announcement
email.

Signed-off-by: Kamal Mostafa <kamal@canonical.com>
Acked-by: Luis Henriques <luis.henriques@canonical.com>