notify-stable-release: new option to edit cover-letter
In some situations it is desirable to edit the cover letter text of a
release.  For example, when sending out the last release of a kernel (i.e.
when a kernel reaches EOL).

This patch adds a new option '--edit' that will allow editing the cover letter
before sending the email.

The code was shamelessly stolen from the notify-stable-review script.

Signed-off-by: Luis Henriques <luis.henriques@canonical.com>