[verify-release-ready] verify the git tag format and value
Verify that the version number in our tag is actually what debian/changelog
says it should be, instead of just a regular expression format match.

Also, if we have a debian/gbp.conf file which specifies the tag format,
use it to validate that our tag is exactly what it should be.

Signed-off-by: Kamal Mostafa <kamal@canonical.com>