apply-stable-patches: fix kernel version identification in 'Cc: stable'
This patch drops the support for identifying the '>', '>=', ... in the
stable kernel versions in the s-o-b area.  This is usually useless: when
the 'Cc: stable' line in a commit contains a kernel version, it is
*always* meant to be interpreted as 'applicable to this kernel version and
above'.

The code being dropped has been responsible for some applicable patches
to have been ignored.  For example, when applying a patch containing the
following (taken from linux commit 97b8ee845393 "ring-buffer: Fix polling
on trace_pipe"):

 Cc: stable@vger.kernel.org # 2.6.27

to stable kernel 3.11, the patch will be identified as not applicable,
because the code assumes it applies *only* to 2.6.27, which is clearly
wrong.

Acked-by: Kamal Mostafa <kamal@canonical.com>
Signed-off-by: Luis Henriques <luis.henriques@canonical.com>