cves-autotriage: handle tag purge rebuilds correctly
When we purge a tag it may well purge dependent tags.  For example when
master and a tag are the same place it is entirely possible the tag
depends on master which will later move.  In this case we need to rescan
both any tags we are trying to generate, and any dependent repos.

Signed-off-by: Andy Whitcroft <apw@canonical.com>