[sb/workflow_engine.py] Eliminate redundant checking of final  conditions before closing the bug.
For all SRU bugs, the packages will _always_ be promoted to -updates.
Most of the time we also promote the packages to -security as well but
no unique tasks are performed to check if that is ready or not.

Since the "check_for_final_close" method is currently performing the check for
both -updates and -security we duplicate the work by calling it twice.

Signed-off-by: Brad Figg <brad.figg@canonical.com>