apply-stable-patches: add option to check commits tagged with 'Fixes:'
This commit adds a new option to also check for commits that have not been
tagged for stable but that contain a 'Fixes:' tag, for a SHA1 that is on
the current branch (both directly or with a 'commit <SHA1> upstream').

Note that this new option does NOT actually apply any patch!  It simply
stores them in a new directory ('Possible-Fixes').  These patches require
extra attention as are likely not suitable for stable trees.

Acked-by: Kamal Mostafa <kamal@canonical.com>
Signed-off-by: Luis Henriques <luis.henriques@canonical.com>