Add support for the 'automated-testing' task
The purpose of the automated-testing task is to track generic testing that
will be done when packages are copied into -proposed.  For the moment,
only the DKMS packages build is handled, but further testing may be added
in the future.

A URL containing the DMKS packages regressions will be polled when the new
task is set to 'Confirmed' (i.e., when packages are copied into -proposed).
The state of the main bug package, the -meta and the -signed packages are
checked for regressions.  If there are regressions in any of these
packages, the task is set to 'Incomplete' which will eventually stall the
bug and manual intervention will be required.

The 'package-testing' task has been removed as it wasn't actually being
used, and thus a lot of dead code is being deleted.

Signed-off-by: Luis Henriques <luis.henriques@canonical.com>