ktl: Debian.changelog(local=False) param
Debian.changelog() and Debian.raw_changelog() now accept a bool param 'local'.

When called with local=False the committed version of the changelog is parsed
(this the default, matching old behavior).

When called with local=True, the local copy of the changelog (possibly locally
modified) is parsed instead.

It may make even more sense to switch the default behavior to local=True but
that's not done here.  This commit introduces no functional change.

Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Kamal Mostafa <kamal@canonical.com>