create-stable-tracker: fix SRU boilerplate list-of-patches text
Make the boilerplate text accurately reflect that the patches listed
below (after we paste them in) are those that still need to be applied,
as opposed to the complete list of patches in the upstream stable
release.  Usually, this is actually just a subset of the upstream stable
release, since we'll have already picked up some of them early.

Signed-off-by: Kamal Mostafa <kamal@canonical.com>