ktl: Don't treat all strings matching "#[0-9]+" as a bug numbers
The bug number regular expression matching correctly identifies
lines containing launchpad bug numbers, but once such a line is
identified it treats any string matching "#[0-9]+" as a bug
number. For example, in this changelog entry #1 is identified as
a bug number:

  * [LTCTest][Opal][OP820] Machine crashed with Oops: Kernel access of bad area,
    sig: 11 [#1] while executing Froze PE Error injection (LP: #1603449)

As far as I can tell the format of bug strings that we need to
match is the string "LP:" followed by whitespace, followed by one
or more bug numbers of the format "#[0-9]+" separated by a comma
and whitespace. This can be expressed as the following python
regex:

  LP:\s#[0-9]+(,\s#[0-9]+)*

I'm not a regex wizard, so there may be a more optimal way to
express this regex, but it seems to work.

Therefore this patch updates the bug number matching as follows:

 1. Use re.finditer with the above regex (modified to use non-
    capturing grouping) to match all groups of bug numbers on a
    line.

 2. Within each match, use findall to create a list of all bug
    numbers by matching against "#([0-9]+)".

 3. Append the list returned by re.findall directly to the bug
    list. The list returned by re.findall will never be empty
    since the string being searched matched the regex used in
    (1), but even if it was appending an empty list to another
    list is a nop.

In my testing this matches the same set of bug numbers matched by
the current code but does not match the #1 in the example above.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kamal Mostafa <kamal@canonical.com>