swm: remove block-proposed tags when all is ready
britney only checks for the tags, not the tasks, and block-proposed will
block it. So, block-proposed tags need to be removed whenever we are
ready to promote the packages.

We have the kernel-block tag as a tool to stop the promotions.

Remove the tags whenever we set any of the release, security or updates
promotion to Confirmed or further.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>