start-sru-cycle: Handle failing bug creation more gracefully
Calling tb.open() can return None when it completely fails to create
the new bug report. One major reason for this seems to be that the
package does not exist on the ubuntu project.
Instead of failing completely, print an error message and go on with
the other backports or derivatives.
Still fail completely if the master bug cannot be created because that
really is ultimately bad.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Kamal Mostafa <kamal@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>