cve-autotriage: switch ROWID to a Tid PRIMARY KEY
It turns out that ROWIDs are inviolate and unchanging _unless_ you vacuum
your database.  If you add a PRIMARY KEY then that will be unchanging
and ROWID will become an alias for it.

Added schema #3 which copies the table to add this new PRIMARY KEY as
you cannot add one once a table is instantiated.

Signed-off-by: Andy Whitcroft <apw@canonical.com>