git-build-kernel: fix series selection for UNRELEASED custom kernels
G-b-k determines the target series and build chroot by examining the
topmost changelog entry -- but if that is in the "UNRELEASED" state,
it falls back to examining the git repo itself (e.g. xenial.git).

This scheme is insufficient for the custom kernels that we've recently
moved to arbitrary repo.git names e.g. linux-something.git -- they can't
figure out their target series when they are the "UNRELEASED" state.

Now, g-b-k will also examine the second changelog entry, if the first is
in the "UNRELEASED" state.

Signed-off-by: Kamal Mostafa <kamal@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Marcelo Cerri <marcelo.cerri@canonical.com>