cve-autotriage: invalidate a branch when parent is invalid
When checking if we have a good scan of a branch check if the parent
tag is valid.  If it is not we are holding that tag live.  Mark
ourselves invalid and rescan.  Once any scan is complete relink anyone
who was depending on our old selves to prevent the rebuild cascading
upwards unnecessarily.  Switch clean to using the same split functions.

Signed-off-by: Andy Whitcroft <apw@canonical.com>