swm: add a routing metaphor to allow mapping of pockets into PPAs for ESM
We currently expect packages for flow from the CKT PPA into the
ubuntu/primary archive passing through Proposed to Updates/Security
(or Release for development packages).

	ppa -> Proposed -> Updates/Security
	ppa -> Proposed -> Release

This same progression exists for ESM but they pass via a set of
PPAs which represent the nominal pockets:

	ppa:		ppa:canonical-kernel-esm/ppa
	Proposed:	ppa:canonical-kernel-esm/proposed
	Security:	ppa:ubuntu-esm/ppa

This commit switches the workflow-manager Pockets to a virtual
contruct.  These are then routed to actual archive/pocket
combinations as appropriate for the routing for this package.
At inception there are two routing modes 'ubuntu/primary' which
represents the existing CKT PPA to ubuntu/primary archive combo,
and ESM which represents the various PPAs above.

Signed-off-by: Andy Whitcroft <apw@canonical.com>