swm: wfl/wft/promoter.py -- apply master checks to devel kernels too
We are correctly collating the master bug requirements based on whether
this is an SRU kernel or a devel one.  We then only initialise the
validator list from that for an SRU kernel.  This leads wfm to blow
up if a devel derivative is found.  As the list is already sanitised
for master type we can use it in all cases.

Signed-off-by: Andy Whitcroft <apw@canonical.com>