swm: promote_to_*.py -- merge _fix_released and _invalid handling
Fix _fix_released and _invalid to better validate security status as
that status does not relate to our status bug to the promote-to-security
and security-signoff tasks.  As this is identicle for both it makes
most sense to merge the two.  Pull up the task oriented handling for
promote-to-security to promote-to-updates and conditionalise the final
message.  Now we have a single common function.  Switch to using that
for both stati.

Signed-off-by: Andy Whitcroft <apw@canonical.com>