DefaultAssigneeMissing: fix usage of message vs msg
Do not use any of msg or message and use str(e) to represent it.

This fixes create-release-tracker or any other use of the Workflow class
when there are no default assignees for a given task. Otherwise, we
would raise an exception for AttributeError and probably bail out of a
script, which happens to create-release-tracker.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>