chroots: clean up chroot package installation
We havea number of issues in the current code to install appropriate
packages to allow test building of kernels.

For no readily apparent reason we are installing most of the build
dependencies for the kernel directly manually, and then request apt
install the build dependencies for the kernel.  This likely was because
we were not originally using --source-only so "apt-get build-dep linux"
would install the build dependencies of linux-meta, basically nothing.
Clean this up by installing build-essential first, then the build
dependencies for linux, then finally any remaining additional packages.
While we are here wire this up to the build_dep configuration entry.

We have a whole heap of build depenedencies which should have been added as
a temporary measure.  When a new build dependency is added we will not have
it in the chroot, but we need to install it to test build the new kernel.
Chicken-and-egg.  These will have to be added but they should not persist
over time.  After the above change I was able to work out which were of
this type and clean them up.

We have some very complex logic to decide which cross compilers to install
into the amd64 chroot.  This is clearly flawed as it tries to install
cross compiler combinations which do not exist and most of these it tries
to install more than once.  As the cross compilers we need are directly
defined by the build architectures of the series, switched this code over
to lookup the (newly added) build_arches configuration for the series.
Based on the architectures therein install appropriate packages.

Cleaned up the additional package installer to directly test whether the
package is alredy installed using dpkg -l.

Cleaned up some old code to install gcc-4.6 and gcc-4.7 to only install
those in series for which those compilers are present.  Also limited
installation to armhf as they are only needed there, or as cross compilers
for armhf.

Signed-off-by: Andy Whitcroft <apw@canonical.com>