maintscripts/verify-release-ready: convert to KernelSeries
Note that we were looking up valid series using the version.  With version
appearing in more than one series with lts backports this is not a valid
approach.  Switched to checking this from series.  The series must exist,
the package must exist in the series, and the version prefix must be a
valid one in that package.

Signed-off-by: Andy Whitcroft <apw@canonical.com>