wfm: snap.py: fix snap store version comparison
The kernel snaps versions are now the same as the source package and not
as the meta package, so the matching of the tracking bug version and the
version on the snap store doesn't work anymore. Use the tracking bug
pkg_version to perform the match, which fixes the issue and is more
precise.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>