start-sru-cycle: change backports / derivatives format in tracking bug description
Now the backports / derivatives in tracking bug description looks like:
backports: 1743363,1743364
derivatives: 1743365,1743367,1743368,1743369,1743370,174337

Add a 'bug' string before the bug number for launchpad to treat it as a link.
Add a '(package-name)' after it, so we could know what it is from a glance.

The new style is expected to be:
backports: bug 1743363 (linux-aws), bug 1743364 (linux-lts-xenial)
derivatives: bug 1743365 (linux-aws), bug 1743367 (linux-euclid), bug 1743368 (linux-gke)...

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kamal Mostafa <kamal@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>