shanky: handle prepare-package* tasks which should not exist
If we have a prepare-package* task which should not have existed we will
crater trying to determine if it is built.  Fix the package so we can
partially instantiate it without a version.  This lets us then determine
which packages we should have and further use that information to
invalidate them early.

Signed-off-by: Andy Whitcroft <apw@canonical.com>