cranky-test-build: Add --commit commandline option
Add a commandline option to specify the commit to use for the test build.
Also create a temporary branch that is pushed to the builder so that
subsequent builds of a multi-arch build always build from the same commit
in case the local repo moves on while the build is in progress.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>