kernel_series: convert copy-forward to a proper dependency
As things stand when copy-forward is specified the meaning
of derived-from changes.  This is utterly confusing, it also
prevents us from defining a kernel such as cosmic/linux-oem as
derived-from bionic/linux but copy-forward from bionic/linux-oem.
Switch copy-forward to be a Dependency object with the same form
as derived-from.

The copy_forward attribute is switched from returning a direct
boolean True/False to returing a valid DependencyEntry or None.
All current consumers of this data do so in a boolean context (and
none directly compare it to True or False).  They will therefore be
unaffected by this change.  The new data reader contains backwards
compatibilty support for the old form; if this field contains a
boolean we will either return the derived-from data or None.  If the
new form data is published and subsequently consumed by the older
code it return either a list when a dependency is found or False.
Again consumers will correctly map this to True/False.

Signed-off-by: Andy Whitcroft <apw@canonical.com>