swm: promote-to-proposed should move Fix Committed when all packages are visible
We are moving promote-to-proposed to Fix Committed when any package is
visible.  However if we do a within ABI bump say to fix a packaging issue
then we will find the existing -meta published and move promote-to-proposed
to Fix Committed when no action has been taken.

Move this to check all packages are in proposed.  Switch any_in_pocket()
to all_in_pocket() to match the semantic change.

Signed-off-by: Andy Whitcroft <apw@canonical.com>