lib/debian: Get the version of the master kernel a changelog refers to
When parsing the changelog of a derivative kernel, there may be a line
indicating the master kernel version that was imported from. The first
occurrence of that line is the one that derivative version should be based on.
Set that in the dictionary that we return when parsing that changelog.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>