swm: bug -- primary package is optional
The primary package is optional for meta only packages.  Only consider
the primary package if we have a valid prepare-package task.  Now that
all packages are optional we must consider and evaluate a package in any
'valid' state else we might think there are no packages and consider
them thus all in -proposed.  Move to considering that we do have a
package if we have a prepare-package task that is not Invalid.

Signed-off-by: Andy Whitcroft <apw@canonical.com>