cranky-update-snap: new script to update snapcraft.yaml
This is a script to automate the update of the snapcraft.yaml for the
kernel snaps. By default it looks for the new source package version on
the -proposed pocket using 'rmadison', updates the yaml file, commits
and tags the commit. Some options are provided to change the default
behavior.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>