cranky: rename cranky.lib to crl
When running python code inside a directory, python will put the program
directory into the python path, preventing some code to import modules from a
namespace named "lib", if there is a "lib" directory there. Also, "lib" as a
namespace is too generic, so rename it to "crl".

The other namespaces should also be renamed, but in the meantime, this alone
allows more code to move to this better naming, and some other changes to
happen under cranky related to module loading.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>