start-sru-cycle: handle derivative chains better
Recently xenial/hwe-edge was re-defined to be derived from
xenial/hwe which in turn is derived from bionic/linux.
This would cause start-sru-cycle to create tracking bugs for
those in the wrong places. Ideally the outcome should be a
xenial/hwe which has a bionic/linux as its master bug but is
in turn master bug for xenial/hwe-edge. However this requires
some larger modifications to the codebase. That will get done
eventually (as part of the work to get start-sru-cycle to be
re-entrant) but unlikely quick enough for next cycle.

So this is a quicker approach of evaluating the oldest ancestor
of a derivative chain. It results in a flattened derivative list
(so with the example above, both xenial/hwe and xenial/hwe-edge
tracking bugs will be created as derivatives of bionix/linux).
Which is the way it was before derivative chains.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>