git-build-kernel: debuild source packages with suitable -vversion
Determine latest released package version in <series>-updates and supply
it to dpkg-genchanges via debuild -vversion.

Special case: If rmadison reports that the version published in -updates
is the same or newer than the source package version being built, then
just do not pass any -v<updates_version> to debuild.  dpkg-genchanges
does actually detect such out-of-order versions but misbehaves if the
updates_version doesn't exist in the changelog at all (as with our
derivative kernels).  This situation only occurs when test-building
source packages which have already been published; the resulting source
package could not be re-uploaded anyway, so the contents of
_source.changes are not critical.

Signed-off-by: Kamal Mostafa <kamal@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Kamal Mostafa <kamal@canonical.com>