cranky: review -- use git for generating diffs
Git can be used to generate diffs for things that aren't git
repositories. This has a couple of nice advantages, namely that
it detects renaming of the abi directories rather than just
showing a bunch of removed and added files. The downside is that
the source packages must be fully extracted, but the benefits
look to be worth it.

Acked-by: Tyler Hicks <tyhicks@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>