cranky-close: add a --base-version parameter
On rare occasions, the changelog won't have the line indicating that changes
have come from a master version. That is the current case for kernels whose
update script just copy previous changelog entries from the master kernel.
Those are going to be changed so they use this tool, which will add them by
calling insert-ubuntu-changes. But the first iteration of those will require
this parameter.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>