cranky-close: do not release changelog when there were no changes
Do not finalize the changelog for a release, and do not commit, if there were
no changes to the tree. The changelog itself would be changed if there were any
changes on the master, or any other tree changes. If there are none of those,
we don't need a new release.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>