ktl/debian: fdr: support dry_run option
Use run_command and takes advantage of its dry_run option when fdr is called
with its own dry_run option as well. As we move from system to run_command, the
output does not go to console anymore, which is intended. But print it in case
the command fails and return its status as well.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>