cranky-start: Really call update-from-*master
When we start the crank, a new commit introducing ABI and config updates from
the master branch should be made after a rebase. The update-from-*master script
does some of these updates, including fixups of config, rules, and what not.
Its changes to the changelog, however, are discarded, and done on a later
stage.

Acked-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>