ktl/tracking_bug2.py: New re-designed tracking bug library
Rewrite the library which is used to work with tracking bugs
and its users (create-release-tracker and start-sru-cycle)
to be more usable.

Main goal is to be able to load a current state from launchpad
so we can either continue from a failed state or automatically
're-link' tracking bugs in case of re-spins.

Acked-by: Andy Whitcroft <apw@canonical.com>
Stigned-off-by: Stefan Bader <stefan.bader@canonical.com>