shanky: promote-to-proposed is gated by master being ready for same
We should allow promote-to-proposed on a derivative bug if the
master is ready to copy.  We check for its promote-to-proposed
being Confirmed which is ready to copy or Fix Released which means
it was. We do not consider states between those which are equally
valid, as someone as approved the thing and copied it.

If anything we should not consider Confirmed but states after that
implying it was approved.

Signed-off-by: Andy Whitcroft <apw@canonical.com>