cranky: crl/handle -- add HandleError to allow simpler error handling
With the first consumers of Handles we are finding they are having to
import a bunch of unexpected error types in order to correctly catch
errors from instantiating bad Handles.

Added a new HandleError type which replaces all of these strange
types.  Convert all internal ValueError types as well so there is
a single type to check for.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>