cranky: crl/config -- handle supplied empty config correctly
While extending the Handle objects I noted during manual testing
that some of the handle tests failed unexpectedly.  This was
occuring because they were loading my personal configuration.
This was occuring when we were supplying empty configuration data
to ensure consistent operation.

Switched to checking data (and filename) to 'is None' forms to
correctly detect the difference between None and ''.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>