cranky: crl/handle -- HandleSet/HandleTree add validate option
In the normal case we wish to ensure that the HandleTree objects
point to trees which are checked out to the series/package required.
Where this is not so we raise an error while instantiating the
HandleTree objects.

For clone we need to instantiate the HandleTree objects against
directories that do not exist, or worse are checked out into another
series/package branch.  In order to add the new remotes etc we need
to indicate we do not care to validate the directory.

Add a validate option which defaults to True.

Acked-by: Seth Forshee <seth.forshee@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>