cranky: test-build -- use the hostname in the git remote URL
It is entirely likley that the hostname for a builder is complex
(kathleen.buildd) and we may not be able to talk to the machine
using the plain shortname (kathleen).  Therefore treat the hostname
as a git remote name, and use the real hostname from the URL that
the remote points us to.

Note that as this also can contain the username@ prefix this removes
the username match restriction.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>