ktl/kernel_series: switch snap track/arches to publish-to
We are publishing some snaps (pi-kernel for example) to multiple tracks
and different track combinations for each architecture.  Add a publish-to
attribute which holds a per architecture list of track names.  To publish
to the default track name use 'latest'.  For compatibility load publish-to
from the existing track/arches elements if not-present.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>