verify-release-ready: fix previous ABI check for backports
We have changed our tooling to extract the ABI files from the packages
also for the backports, instead of copying them from the master kernel.
And also to always create the ABI directory with the full version of the
source package.

We reverted the change we made on 'final-checks' to remove the trailing
backport suffix, so it's now safe to look for the ABI files on the
full-version directory name.

Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>