cranky-start: pass HOME env to git subshell
Depending on the environment, the git that is run on a subshell to
re-order the start new release commit is not able to find the .gitconfig
file on the user home dir, which causes an error when re-applying the
commits. Fix it by passing down the $HOME environment variable to the
subshell.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>