config-review: change top-level Kconfig file
The top-level Kconfig file is Kconfig, not arch/<arch>/Kconfig.
Not sure how long this has been the case, but in 4.19 many
Kconfig files are not included when starting from the arch
Kconfig. Update the tooling to start at Kconfig instead and pass
in the arch so we can expand the SRCARCH variable.

This may break older kernels, but it probably doesn't matter as
these scripts never seem to be used outside of devel. If it does
turn out to be a problem we can add some version checking to
restore the previous behavior for older kernels.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>