ktl/announce: split out the command interface for announce.py
We cannot get the imports right for direct use of announce.py as
a command which breaks us importing any sub-functions from ktl in
announce.py itself.

Rip out the primary interface and expose it from bin/announce.
Let that use ktl.announce from outside as it should.  Convert all
of the users of the direct interface.

Signed-off-by: Andy Whitcroft <apw@canonical.com>